chore: moved LogProbParser, transformer mixin, in parser.py
